psbt: move output into new file
